django_project/requirements.txt
asgiref==3.4.1
boto3==1.24.36
botocore==1.27.36
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.10
whitenoise==6.2.0
